+++ modules/spring-web/src/main/kotlin/com/explyt/spring/web/builder/openapi/OpenApiServerBuilder.kt
@@ -19,14 +19,5 @@ package com.explyt.spring.web.builder.openapi
 
 import com.explyt.spring.web.builder.AbstractBuilder
 
-abstract class OpenApiPathHttpTypeBuilder(indent: String, builder: StringBuilder) : AbstractBuilder(indent, builder) {
-
-    protected fun contentTypes(types: Collection<String>): List<String> {
-        return if (types.isEmpty()) {
-            listOf("application/json")
-        } else {
-            types.toList()
-        }
-    }
-
-}
+abstract class OpenApiServerBuilder(indent: String, builder: StringBuilder) :
+    AbstractBuilder(indent, builder)