+++ modules/spring-web/src/main/kotlin/com/explyt/spring/web/inspections/RequestMappingDuplicateInspection.kt
@@ -24,23 +24,100 @@ import com.explyt.spring.web.SpringWebClasses
 import com.explyt.util.ExplytPsiUtil.getHighlightRange
 import com.explyt.util.ExplytPsiUtil.isMetaAnnotatedBy
 import com.explyt.util.ExplytPsiUtil.isMetaAnnotatedByOrSelf
+import com.explyt.util.ExplytPsiUtil.isPublic
 import com.intellij.codeInsight.AnnotationUtil
+import com.intellij.codeInspection.InspectionManager
+import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.openapi.module.ModuleUtilCore
-import com.intellij.psi.PsiElementVisitor
+import com.intellij.psi.PsiAnnotationMemberValue
 import com.intellij.psi.PsiMethod
-import com.intellij.uast.UastVisitorAdapter
+import io.ktor.http.*
 import org.jetbrains.kotlin.build.joinToReadableString
 import org.jetbrains.uast.*
 import org.jetbrains.uast.visitor.AbstractUastNonRecursiveVisitor
 
 class RequestMappingDuplicateInspection : SpringBaseUastLocalInspectionTool() {
 
-    override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
-        return UastVisitorAdapter(RequestMappingVisitor(holder, isOnTheFly), true)
+    override fun checkClass(
+        aClass: UClass, manager: InspectionManager, isOnTheFly: Boolean
+    ): Array<ProblemDescriptor> {
+        val javaPsi = aClass.javaPsi
+        if (!javaPsi.isMetaAnnotatedBy(SpringWebClasses.CONTROLLER)
+            && !javaPsi.isMetaAnnotatedBy(SpringWebClasses.REQUEST_MAPPING)
+        ) return emptyArray()
+        val module = ModuleUtilCore.findModuleForPsiElement(javaPsi) ?: return emptyArray()
+        val requestMappingMah = MetaAnnotationsHolder.of(module, SpringWebClasses.REQUEST_MAPPING)
+
+        val methodsByUrlMap = javaPsi.methods.asSequence()
+            .filter { it.isPublic }
+            .filter { it.isMetaAnnotatedBy(SpringWebClasses.REQUEST_MAPPING) }
+            .flatMap { toUrlMethodPair(requestMappingMah, it) }
+            .groupBy { it.methodType + it.url }
+
+        val holder = ProblemsHolder(manager, javaPsi.containingFile, isOnTheFly)
+        registerProblems(methodsByUrlMap, holder, manager, isOnTheFly)
+        return holder.resultsArray
     }
 
+    private fun registerProblems(
+        methodsByUrlMap: Map<String, List<DuplicateUrlInfo>>,
+        holder: ProblemsHolder,
+        manager: InspectionManager,
+        isOnTheFly: Boolean
+    ) {
+        methodsByUrlMap.forEach { (_, duplicates) ->
+            if (duplicates.size < 2) return@forEach
+            for (duplicate in duplicates) {
+                val psiElement = duplicate.psiAnnotationMemberValue.toUElement()?.sourcePsi ?: continue
+                holder.registerProblem(
+                    manager.createProblemDescriptor(
+                        psiElement, psiElement.getHighlightRange(),
+                        SpringWebBundle.message(
+                            "explyt.spring.web.inspection.requestMapping.duplicate.methods",
+                            duplicates.filter { it.psiMethod.name != duplicate.psiMethod.name }
+                                .map { it.psiMethod.name }.joinToReadableString()
+                        ),
+                        ProblemHighlightType.GENERIC_ERROR,
+                        isOnTheFly
+                    )
+                )
+            }
+        }
+    }
+
+    private fun toUrlMethodPair(
+        requestMappingMah: MetaAnnotationsHolder, psiMethod: PsiMethod
+    ): List<DuplicateUrlInfo> {
+        return requestMappingMah.getAnnotationMemberValues(psiMethod, TARGET_VALUE)
+            .flatMap { toDuplicateUrlInfo(it, psiMethod, requestMappingMah) }
+    }
+
+    private fun toDuplicateUrlInfo(
+        psiValue: PsiAnnotationMemberValue, psiMethod: PsiMethod, requestMappingMah: MetaAnnotationsHolder
+    ): List<DuplicateUrlInfo> {
+        val urlString = AnnotationUtil.getStringAttributeValue(psiValue) ?: return emptyList()
+        val httpMethods = requestMappingMah.getHttpMethods(psiMethod)
+        return httpMethods.map { DuplicateUrlInfo(Url(urlString), it, psiMethod, psiValue) }
+    }
+
+    private fun MetaAnnotationsHolder.getHttpMethods(
+        psiMethod: PsiMethod
+    ) = getAnnotationMemberValues(psiMethod, TARGET_METHOD)
+        .mapTo(mutableSetOf()) { it.text.split('.').last() }
+
+    companion object {
+        private val TARGET_VALUE = setOf("value")
+        private val TARGET_METHOD = setOf("method")
+    }
+
+    private data class DuplicateUrlInfo(
+        val url: Url,
+        val methodType: String,
+        val psiMethod: PsiMethod,
+        val psiAnnotationMemberValue: PsiAnnotationMemberValue,
+    )
 }
 
 private class RequestMappingVisitor(